Fix JavaScript bundle TypeError
The constants exported from each of `alert.js`, `btn.js`, and
`dropdown.js` were newly created references to `jQuery` objects.
Each module file invoked `jQuery` in its callback form requesting that
the function run at document ready. The modules then saved off a
reference to the return value and exported it as a constant.

The TypeError seen in the console was from attempting to invoke these
new constants as if they were functions:

    const foo = jQuery(function() {});
    foo();
    // Uncaught TypeError: foo is not a function

This patch proposes that modules simply export the functions by
themselves. Then, consumers of the modules are free to do whatever
they wish with the functions. Maybe a test can import the function.
Maybe a consumer of the module can invoke the function whenever
it is ready.

Here I attempt to preserve the previous behavior; that each function
is scheduled to run at document ready.

_cf._ jQuery(callback) https://api.jquery.com/jQuery/#jQuery3